Refactor mutlipartuploaderserice to include logic for all uploads
