+++ src/OctoshiftCLI.IntegrationTests/BbsToGithub.cs
@@ -28,7 +28,7 @@ public sealed class BbsToGithub : IDisposable
     private readonly HttpClient _sourceBbsHttpClient;
     private readonly BbsClient _sourceBbsClient;
     private readonly BlobServiceClient _blobServiceClient;
-    private readonly MultipartUploaderService _multipartUploader;
+    private readonly ArchiveUploader _multipartUploader;
     private readonly Dictionary<string, string> _tokens;
     private readonly DateTime _startTime;
     private readonly string _azureStorageConnectionString;
@@ -58,7 +58,7 @@ public sealed class BbsToGithub : IDisposable
 
         _targetGithubHttpClient = new HttpClient();
         _targetGithubClient = new GithubClient(_logger, _targetGithubHttpClient, new VersionChecker(_versionClient, _logger), new RetryPolicy(_logger), new DateTimeProvider(), targetGithubToken);
-        _multipartUploader = new MultipartUploaderService(_targetGithubClient, _logger);
+        _multipartUploader = new ArchiveUploader(_targetGithubClient, _logger);
         _targetGithubApi = new GithubApi(_targetGithubClient, "https://api.github.com", new RetryPolicy(_logger), _multipartUploader);
 
         _blobServiceClient = new BlobServiceClient(_azureStorageConnectionString);
