+++ src/OctoshiftCLI.IntegrationTests/GithubToGithub.cs
@@ -21,7 +21,7 @@ namespace OctoshiftCLI.IntegrationTests
         private bool disposedValue;
         private readonly Dictionary<string, string> _tokens;
         private readonly DateTime _startTime;
-        private readonly MultipartUploaderService _multipartUploader;
+        private readonly ArchiveUploader _multipartUploader;
 
         public GithubToGithub(ITestOutputHelper output)
         {
@@ -36,7 +36,7 @@ namespace OctoshiftCLI.IntegrationTests
             _githubHttpClient = new HttpClient();
             _versionClient = new HttpClient();
             _githubClient = new GithubClient(logger, _githubHttpClient, new VersionChecker(_versionClient, logger), new RetryPolicy(logger), new DateTimeProvider(), githubToken);
-            _multipartUploader = new MultipartUploaderService(_githubClient, logger);
+            _multipartUploader = new ArchiveUploader(_githubClient, logger);
             _githubApi = new GithubApi(_githubClient, "https://api.github.com", new RetryPolicy(logger), _multipartUploader);
 
             _helper = new TestHelper(_output, _githubApi, _githubClient);
