+++ src/OctoshiftCLI.Tests/Octoshift/Services/GithubApiTests.cs
@@ -21,7 +21,7 @@ public class GithubApiTests
     private const string API_URL = "https://api.github.com";
     private readonly RetryPolicy _retryPolicy = new(TestHelpers.CreateMock<OctoLogger>().Object) { _httpRetryInterval = 0, _retryInterval = 0 };
     private readonly Mock<GithubClient> _githubClientMock = TestHelpers.CreateMock<GithubClient>();
-    private readonly MultipartUploaderService _multipartUploader;
+    private readonly ArchiveUploader _multipartUploader;
 
     private readonly GithubApi _githubApi;
 
@@ -48,7 +48,7 @@ public class GithubApiTests
     public GithubApiTests()
     {
         var logger = new OctoLogger(_ => { }, _ => { }, _ => { }, _ => { });
-        _multipartUploader = new MultipartUploaderService(_githubClientMock.Object, logger);
+        _multipartUploader = new ArchiveUploader(_githubClientMock.Object, logger);
         _githubApi = new GithubApi(_githubClientMock.Object, API_URL, _retryPolicy, _multipartUploader);
     }
 
@@ -428,7 +428,7 @@ public class GithubApiTests
 
         // Act
         var logger = new OctoLogger(_ => { }, _ => { }, _ => { }, _ => { });
-        var multipartUploader = new MultipartUploaderService(_githubClientMock.Object, logger);
+        var multipartUploader = new ArchiveUploader(_githubClientMock.Object, logger);
         var githubApi = new GithubApi(_githubClientMock.Object, API_URL, _retryPolicy, multipartUploader);
         await githubApi.RemoveTeamMember(GITHUB_ORG, teamName, member);
 
@@ -451,7 +451,7 @@ public class GithubApiTests
 
         // Act
         var logger = TestHelpers.CreateMock<OctoLogger>().Object;
-        var multipartUploader = new MultipartUploaderService(_githubClientMock.Object, logger);
+        var multipartUploader = new ArchiveUploader(_githubClientMock.Object, logger);
         var githubApi = new GithubApi(_githubClientMock.Object, API_URL, _retryPolicy, multipartUploader);
         await githubApi.RemoveTeamMember(GITHUB_ORG, teamName, member);
