+++ src/bbs2gh/Factories/GithubApiFactory.cs
@@ -30,7 +30,7 @@ public class GithubApiFactory : ITargetGithubApiFactory
         apiUrl ??= DEFAULT_API_URL;
         targetPersonalAccessToken ??= _environmentVariableProvider.TargetGithubPersonalAccessToken();
         var githubClient = new GithubClient(_octoLogger, _client, _versionProvider, _retryPolicy, _dateTimeProvider, targetPersonalAccessToken);
-        var multipartUploader = new MultipartUploaderService(githubClient, _octoLogger);
+        var multipartUploader = new ArchiveUploader(githubClient, _octoLogger);
         return new GithubApi(githubClient, apiUrl, _retryPolicy, multipartUploader);
     }
 }
